Tests was added

+++ ObjectPrintingShould/ObjectsForTest/Collections.cs
@@ -6,5 +6,5 @@ public class Collections
 {
     public int[] Arr { get; set; }
     public List<string> Lis { get; set; }
-    public Dictionary<int, double> Dict { get; set; }
+    public Dictionary<int, int> Dict { get; set; }
 }
